# tests: Replace `FreshSchema` architecture with `TestDatabase`