--- a/src/tests/dump_db.rs
+++ b/src/tests/dump_db.rs
@@ -1,21 +1,21 @@
-use crate::util::FreshSchema;
+use crate::util::TestDatabase;
 use crates_io::worker::dump_db;
 
 #[test]
 fn dump_db_and_reimport_dump() {
     crates_io::util::tracing::init_for_test();
 
-    let database_url = crate::env("TEST_DATABASE_URL");
+    let db_one = TestDatabase::new();
 
     // TODO prefill database with some data
 
     let directory = dump_db::DumpDirectory::create().unwrap();
-    directory.populate(&database_url).unwrap();
+    directory.populate(db_one.url()).unwrap();
 
-    let schema = FreshSchema::new(&database_url);
+    let db_two = TestDatabase::new();
 
     let import_script = directory.export_dir.join("import.sql");
-    dump_db::run_psql(&import_script, schema.database_url()).unwrap();
+    dump_db::run_psql(&import_script, db_two.url()).unwrap();
 
     // TODO: Consistency checks on the re-imported data?
 }