### Before (`FreshSchema`)

The `FreshSchema::new()` fn was creating a new Postgres schema in the test database, and then running the migrations on that new schema to create the tables, indices, routines, etc.

Every test using `FreshSchema::new()` was spending about 180ms on connecting to the database, only 1ms to create the new schema, and then about 230ms on running the migrations.

Most of our tests are currently using https://github.com/allaboutapps/integresql#approach-2a-isolation-by-transactions instead of `FreshSchema::new()`, so this change is currently only relevant for a small minority of our test suite.

### After (`TestDatabase`)

Instead of creating new schemas for each test, `TestDatabase::new()` creates completely new databases for each test. This sounds slower at first glance, and with a naive implementation it certainly is. Creating a database takes around 60-70ms, while a new schema can be created in just 1ms.

The advantage of creating a new database is that they can be created based on template databases. This PR implements it so that on first request a template database is created, the migrations applied and for any subsequent request that template database is used to create new databases. Additionally, the corresponding `TemplateDatabase` struct keeps a pool of database management connections open to avoid the delay of first having to connect to the database.

The creation of the template database takes around 500ms, but then every subsequent test database creation is only taking around 70ms, compared to the ~400ms for `FreshSchema::new()`.

This is roughly based upon https://github.com/allaboutapps/integresql#approach-3a-isolation-by-templates, and that document is giving pointers on how the performance could be improved even further.

### TL;DR

We trade a slightly higher startup cost for a significant speed-up per test.